Fix naming, docs and remove unnecessary types
